pyproject.toml
[project]
name = "debitor"
version = "0.1.0"
description = "Voice agent on LangGraph"
authors = [{ name = "ACME", email = "dev@acme.io" }]
requires-python = ">=3.10"
readme = "README.md"

# Базовые зависимости рантайма
dependencies = [
  "langgraph>=0.2",
  "langchain>=0.3",
  "pydantic>=2.7",
  "python-dotenv>=1.0",
  "loguru>=0.7",
  "streamlit>=1.48.0",
  "typer>=0.16.0",
  "langgraph-checkpoint-sqlite>=2.0.11",
  "langchain-openai>=0.3.29",
  "opentelemetry-api>=1.36.0",
  "opentelemetry-sdk>=1.36.0",
  "opentelemetry-exporter-otlp>=1.36.0",
  "opentelemetry-instrumentation-fastapi>=0.57b0",
  "opentelemetry-instrumentation-httpx>=0.57b0",
  "opentelemetry-instrumentation-logging>=0.57b0",
  "pydantic-settings>=2.10.1",
]

[dependency-groups]
# dev-группа собирается/устанавливается: `uv sync --group dev`
dev = [
  "ruff>=0.5",
  "pre-commit>=3.7"
]

[tool.mypy]
python_version = "3.10"
strict = true
warn_unused_ignores = true

[tool.ruff]
line-length = 100
select = ["E","F","I","B","UP","C90","N","S","W","PL"]
ignore = ["E501"]

[tool.setuptools]
packages = [
  "src"
]
